chore: move session info assembly to utils
